requirements.txt
requests
beautifulsoup4 
torch 
torchvision 
transformers 
pillow
python-whois
pandas
numpy
scikit-learn
opencv-python
flask
tensorflow
imblearn
